refactor: module import update for assertion
